LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/animation/flex-basis-interpolation-expected.txt


FAIL CSS Transitions: property <flex-basis> from neutral to [2%] at (-0.3) should be [0.7%] assert_equals: expected "0.7 % " but got "2 % "
FAIL CSS Transitions: property <flex-basis> from neutral to [2%] at (0) should be [1%] assert_equals: expected "1 % " but got "2 % "
FAIL CSS Transitions: property <flex-basis> from neutral to [2%] at (0.3) should be [1.3%] assert_equals: expected "1.3 % " but got "2 % "
FAIL CSS Transitions: property <flex-basis> from neutral to [2%] at (0.6) should be [1.6%] assert_equals: expected "1.6 % " but got "2 % "
PASS CSS Transitions: property <flex-basis> from neutral to [2%] at (1) should be [2%]
FAIL CSS Transitions: property <flex-basis> from neutral to [2%] at (1.5) should be [2.5%] assert_equals: expected "2.5 % " but got "2 % "
PASS CSS Transitions with transition: all: property <flex-basis> from neutral to [2%] at (-0.3) should be [0.7%]
PASS CSS Transitions with transition: all: property <flex-basis> from neutral to [2%] at (0) should be [1%]
PASS CSS Transitions with transition: all: property <flex-basis> from neutral to [2%] at (0.3) should be [1.3%]
PASS CSS Transitions with transition: all: property <flex-basis> from neutral to [2%] at (0.6) should be [1.6%]
PASS CSS Transitions with transition: all: property <flex-basis> from neutral to [2%] at (1) should be [2%]
PASS CSS Transitions with transition: all: property <flex-basis> from neutral to [2%] at (1.5) should be [2.5%]
FAIL CSS Animations: property <flex-basis> from neutral to [2%] at (-0.3) should be [0.7%] assert_equals: expected "0.7 % " but got "1 % "
PASS CSS Animations: property <flex-basis> from neutral to [2%] at (0) should be [1%]
FAIL CSS Animations: property <flex-basis> from neutral to [2%] at (0.3) should be [1.3%] assert_equals: expected "1.3 % " but got "1 % "
FAIL CSS Animations: property <flex-basis> from neutral to [2%] at (0.6) should be [1.6%] assert_equals: expected "1.6 % " but got "1 % "
FAIL CSS Animations: property <flex-basis> from neutral to [2%] at (1) should be [2%] assert_equals: expected "2 % " but got "1 % "
FAIL CSS Animations: property <flex-basis> from neutral to [2%] at (1.5) should be [2.5%] assert_equals: expected "2.5 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from neutral to [2%] at (-0.3) should be [0.7%] assert_equals: expected "0.7 % " but got "1 % "
PASS Web Animations: property <flex-basis> from neutral to [2%] at (0) should be [1%]
FAIL Web Animations: property <flex-basis> from neutral to [2%] at (0.3) should be [1.3%] assert_equals: expected "1.3 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from neutral to [2%] at (0.6) should be [1.6%] assert_equals: expected "1.6 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from neutral to [2%] at (1) should be [2%] assert_equals: expected "2 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from neutral to [2%] at (1.5) should be [2.5%] assert_equals: expected "2.5 % " but got "1 % "
PASS CSS Transitions: property <flex-basis> from [initial] to [2%] at (-0.3) should be [2%]
PASS CSS Transitions: property <flex-basis> from [initial] to [2%] at (0) should be [2%]
PASS CSS Transitions: property <flex-basis> from [initial] to [2%] at (0.3) should be [2%]
PASS CSS Transitions: property <flex-basis> from [initial] to [2%] at (0.5) should be [2%]
PASS CSS Transitions: property <flex-basis> from [initial] to [2%] at (0.6) should be [2%]
PASS CSS Transitions: property <flex-basis> from [initial] to [2%] at (1) should be [2%]
PASS CSS Transitions: property <flex-basis> from [initial] to [2%] at (1.5) should be [2%]
FAIL CSS Transitions with transition: all: property <flex-basis> from [initial] to [2%] at (-0.3) should be [2%] assert_equals: expected "2 % " but got "auto "
FAIL CSS Transitions with transition: all: property <flex-basis> from [initial] to [2%] at (0) should be [2%] assert_equals: expected "2 % " but got "auto "
FAIL CSS Transitions with transition: all: property <flex-basis> from [initial] to [2%] at (0.3) should be [2%] assert_equals: expected "2 % " but got "auto "
PASS CSS Transitions with transition: all: property <flex-basis> from [initial] to [2%] at (0.5) should be [2%]
PASS CSS Transitions with transition: all: property <flex-basis> from [initial] to [2%] at (0.6) should be [2%]
PASS CSS Transitions with transition: all: property <flex-basis> from [initial] to [2%] at (1) should be [2%]
PASS CSS Transitions with transition: all: property <flex-basis> from [initial] to [2%] at (1.5) should be [2%]
FAIL CSS Animations: property <flex-basis> from [initial] to [2%] at (-0.3) should be [initial] assert_equals: expected "auto " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [initial] to [2%] at (0) should be [initial] assert_equals: expected "auto " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [initial] to [2%] at (0.3) should be [initial] assert_equals: expected "auto " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [initial] to [2%] at (0.5) should be [2%] assert_equals: expected "2 % " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [initial] to [2%] at (0.6) should be [2%] assert_equals: expected "2 % " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [initial] to [2%] at (1) should be [2%] assert_equals: expected "2 % " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [initial] to [2%] at (1.5) should be [2%] assert_equals: expected "2 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from [initial] to [2%] at (-0.3) should be [initial] assert_equals: expected "auto " but got "1 % "
FAIL Web Animations: property <flex-basis> from [initial] to [2%] at (0) should be [initial] assert_equals: expected "auto " but got "1 % "
FAIL Web Animations: property <flex-basis> from [initial] to [2%] at (0.3) should be [initial] assert_equals: expected "auto " but got "1 % "
FAIL Web Animations: property <flex-basis> from [initial] to [2%] at (0.5) should be [2%] assert_equals: expected "2 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from [initial] to [2%] at (0.6) should be [2%] assert_equals: expected "2 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from [initial] to [2%] at (1) should be [2%] assert_equals: expected "2 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from [initial] to [2%] at (1.5) should be [2%] assert_equals: expected "2 % " but got "1 % "
FAIL CSS Transitions: property <flex-basis> from [inherit] to [2%] at (-0.3) should be [3.3%] assert_equals: expected "3.3 % " but got "2 % "
FAIL CSS Transitions: property <flex-basis> from [inherit] to [2%] at (0) should be [3%] assert_equals: expected "3 % " but got "2 % "
FAIL CSS Transitions: property <flex-basis> from [inherit] to [2%] at (0.3) should be [2.7%] assert_equals: expected "2.7 % " but got "2 % "
FAIL CSS Transitions: property <flex-basis> from [inherit] to [2%] at (0.6) should be [2.4%] assert_equals: expected "2.4 % " but got "2 % "
PASS CSS Transitions: property <flex-basis> from [inherit] to [2%] at (1) should be [2%]
FAIL CSS Transitions: property <flex-basis> from [inherit] to [2%] at (1.5) should be [1.5%] assert_equals: expected "1.5 % " but got "2 % "
PASS CSS Transitions with transition: all: property <flex-basis> from [inherit] to [2%] at (-0.3) should be [3.3%]
PASS CSS Transitions with transition: all: property <flex-basis> from [inherit] to [2%] at (0) should be [3%]
PASS CSS Transitions with transition: all: property <flex-basis> from [inherit] to [2%] at (0.3) should be [2.7%]
PASS CSS Transitions with transition: all: property <flex-basis> from [inherit] to [2%] at (0.6) should be [2.4%]
PASS CSS Transitions with transition: all: property <flex-basis> from [inherit] to [2%] at (1) should be [2%]
PASS CSS Transitions with transition: all: property <flex-basis> from [inherit] to [2%] at (1.5) should be [1.5%]
FAIL CSS Animations: property <flex-basis> from [inherit] to [2%] at (-0.3) should be [3.3%] assert_equals: expected "3.3 % " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [inherit] to [2%] at (0) should be [3%] assert_equals: expected "3 % " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [inherit] to [2%] at (0.3) should be [2.7%] assert_equals: expected "2.7 % " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [inherit] to [2%] at (0.6) should be [2.4%] assert_equals: expected "2.4 % " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [inherit] to [2%] at (1) should be [2%] assert_equals: expected "2 % " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [inherit] to [2%] at (1.5) should be [1.5%] assert_equals: expected "1.5 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from [inherit] to [2%] at (-0.3) should be [3.3%] assert_equals: expected "3.3 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from [inherit] to [2%] at (0) should be [3%] assert_equals: expected "3 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from [inherit] to [2%] at (0.3) should be [2.7%] assert_equals: expected "2.7 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from [inherit] to [2%] at (0.6) should be [2.4%] assert_equals: expected "2.4 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from [inherit] to [2%] at (1) should be [2%] assert_equals: expected "2 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from [inherit] to [2%] at (1.5) should be [1.5%] assert_equals: expected "1.5 % " but got "1 % "
PASS CSS Transitions: property <flex-basis> from [unset] to [2%] at (-0.3) should be [2%]
PASS CSS Transitions: property <flex-basis> from [unset] to [2%] at (0) should be [2%]
PASS CSS Transitions: property <flex-basis> from [unset] to [2%] at (0.3) should be [2%]
PASS CSS Transitions: property <flex-basis> from [unset] to [2%] at (0.5) should be [2%]
PASS CSS Transitions: property <flex-basis> from [unset] to [2%] at (0.6) should be [2%]
PASS CSS Transitions: property <flex-basis> from [unset] to [2%] at (1) should be [2%]
PASS CSS Transitions: property <flex-basis> from [unset] to [2%] at (1.5) should be [2%]
FAIL CSS Transitions with transition: all: property <flex-basis> from [unset] to [2%] at (-0.3) should be [2%] assert_equals: expected "2 % " but got "auto "
FAIL CSS Transitions with transition: all: property <flex-basis> from [unset] to [2%] at (0) should be [2%] assert_equals: expected "2 % " but got "auto "
FAIL CSS Transitions with transition: all: property <flex-basis> from [unset] to [2%] at (0.3) should be [2%] assert_equals: expected "2 % " but got "auto "
PASS CSS Transitions with transition: all: property <flex-basis> from [unset] to [2%] at (0.5) should be [2%]
PASS CSS Transitions with transition: all: property <flex-basis> from [unset] to [2%] at (0.6) should be [2%]
PASS CSS Transitions with transition: all: property <flex-basis> from [unset] to [2%] at (1) should be [2%]
PASS CSS Transitions with transition: all: property <flex-basis> from [unset] to [2%] at (1.5) should be [2%]
FAIL CSS Animations: property <flex-basis> from [unset] to [2%] at (-0.3) should be [unset] assert_equals: expected "auto " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [unset] to [2%] at (0) should be [unset] assert_equals: expected "auto " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [unset] to [2%] at (0.3) should be [unset] assert_equals: expected "auto " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [unset] to [2%] at (0.5) should be [2%] assert_equals: expected "2 % " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [unset] to [2%] at (0.6) should be [2%] assert_equals: expected "2 % " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [unset] to [2%] at (1) should be [2%] assert_equals: expected "2 % " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [unset] to [2%] at (1.5) should be [2%] assert_equals: expected "2 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from [unset] to [2%] at (-0.3) should be [unset] assert_equals: expected "auto " but got "1 % "
FAIL Web Animations: property <flex-basis> from [unset] to [2%] at (0) should be [unset] assert_equals: expected "auto " but got "1 % "
FAIL Web Animations: property <flex-basis> from [unset] to [2%] at (0.3) should be [unset] assert_equals: expected "auto " but got "1 % "
FAIL Web Animations: property <flex-basis> from [unset] to [2%] at (0.5) should be [2%] assert_equals: expected "2 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from [unset] to [2%] at (0.6) should be [2%] assert_equals: expected "2 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from [unset] to [2%] at (1) should be [2%] assert_equals: expected "2 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from [unset] to [2%] at (1.5) should be [2%] assert_equals: expected "2 % " but got "1 % "
FAIL CSS Transitions: property <flex-basis> from [0px] to [100px] at (-0.3) should be [0px] assert_equals: expected "0px " but got "100px "
FAIL CSS Transitions: property <flex-basis> from [0px] to [100px] at (0) should be [0px] assert_equals: expected "0px " but got "100px "
FAIL CSS Transitions: property <flex-basis> from [0px] to [100px] at (0.4) should be [40px] assert_equals: expected "40px " but got "100px "
FAIL CSS Transitions: property <flex-basis> from [0px] to [100px] at (0.6) should be [60px] assert_equals: expected "60px " but got "100px "
PASS CSS Transitions: property <flex-basis> from [0px] to [100px] at (1) should be [100px]
FAIL CSS Transitions: property <flex-basis> from [0px] to [100px] at (1.5) should be [150px] assert_equals: expected "150px " but got "100px "
FAIL CSS Transitions with transition: all: property <flex-basis> from [0px] to [100px] at (-0.3) should be [0px] assert_equals: expected "0px " but got "- 30px "
PASS CSS Transitions with transition: all: property <flex-basis> from [0px] to [100px] at (0) should be [0px]
PASS CSS Transitions with transition: all: property <flex-basis> from [0px] to [100px] at (0.4) should be [40px]
PASS CSS Transitions with transition: all: property <flex-basis> from [0px] to [100px] at (0.6) should be [60px]
PASS CSS Transitions with transition: all: property <flex-basis> from [0px] to [100px] at (1) should be [100px]
PASS CSS Transitions with transition: all: property <flex-basis> from [0px] to [100px] at (1.5) should be [150px]
FAIL CSS Animations: property <flex-basis> from [0px] to [100px] at (-0.3) should be [0px] assert_equals: expected "0px " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [0px] to [100px] at (0) should be [0px] assert_equals: expected "0px " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [0px] to [100px] at (0.4) should be [40px] assert_equals: expected "40px " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [0px] to [100px] at (0.6) should be [60px] assert_equals: expected "60px " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [0px] to [100px] at (1) should be [100px] assert_equals: expected "100px " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [0px] to [100px] at (1.5) should be [150px] assert_equals: expected "150px " but got "1 % "
FAIL Web Animations: property <flex-basis> from [0px] to [100px] at (-0.3) should be [0px] assert_equals: expected "0px " but got "1 % "
FAIL Web Animations: property <flex-basis> from [0px] to [100px] at (0) should be [0px] assert_equals: expected "0px " but got "1 % "
FAIL Web Animations: property <flex-basis> from [0px] to [100px] at (0.4) should be [40px] assert_equals: expected "40px " but got "1 % "
FAIL Web Animations: property <flex-basis> from [0px] to [100px] at (0.6) should be [60px] assert_equals: expected "60px " but got "1 % "
FAIL Web Animations: property <flex-basis> from [0px] to [100px] at (1) should be [100px] assert_equals: expected "100px " but got "1 % "
FAIL Web Animations: property <flex-basis> from [0px] to [100px] at (1.5) should be [150px] assert_equals: expected "150px " but got "1 % "
FAIL CSS Transitions: property <flex-basis> from [0%] to [100%] at (-0.3) should be [0%] assert_equals: expected "0 % " but got "100 % "
FAIL CSS Transitions: property <flex-basis> from [0%] to [100%] at (0) should be [0%] assert_equals: expected "0 % " but got "100 % "
FAIL CSS Transitions: property <flex-basis> from [0%] to [100%] at (0.4) should be [40%] assert_equals: expected "40 % " but got "100 % "
FAIL CSS Transitions: property <flex-basis> from [0%] to [100%] at (0.6) should be [60%] assert_equals: expected "60 % " but got "100 % "
PASS CSS Transitions: property <flex-basis> from [0%] to [100%] at (1) should be [100%]
FAIL CSS Transitions: property <flex-basis> from [0%] to [100%] at (1.5) should be [150%] assert_equals: expected "150 % " but got "100 % "
FAIL CSS Transitions with transition: all: property <flex-basis> from [0%] to [100%] at (-0.3) should be [0%] assert_equals: expected "0 % " but got "- 30 % "
PASS CSS Transitions with transition: all: property <flex-basis> from [0%] to [100%] at (0) should be [0%]
PASS CSS Transitions with transition: all: property <flex-basis> from [0%] to [100%] at (0.4) should be [40%]
PASS CSS Transitions with transition: all: property <flex-basis> from [0%] to [100%] at (0.6) should be [60%]
PASS CSS Transitions with transition: all: property <flex-basis> from [0%] to [100%] at (1) should be [100%]
PASS CSS Transitions with transition: all: property <flex-basis> from [0%] to [100%] at (1.5) should be [150%]
FAIL CSS Animations: property <flex-basis> from [0%] to [100%] at (-0.3) should be [0%] assert_equals: expected "0 % " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [0%] to [100%] at (0) should be [0%] assert_equals: expected "0 % " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [0%] to [100%] at (0.4) should be [40%] assert_equals: expected "40 % " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [0%] to [100%] at (0.6) should be [60%] assert_equals: expected "60 % " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [0%] to [100%] at (1) should be [100%] assert_equals: expected "100 % " but got "1 % "
FAIL CSS Animations: property <flex-basis> from [0%] to [100%] at (1.5) should be [150%] assert_equals: expected "150 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from [0%] to [100%] at (-0.3) should be [0%] assert_equals: expected "0 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from [0%] to [100%] at (0) should be [0%] assert_equals: expected "0 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from [0%] to [100%] at (0.4) should be [40%] assert_equals: expected "40 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from [0%] to [100%] at (0.6) should be [60%] assert_equals: expected "60 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from [0%] to [100%] at (1) should be [100%] assert_equals: expected "100 % " but got "1 % "
FAIL Web Animations: property <flex-basis> from [0%] to [100%] at (1.5) should be [150%] assert_equals: expected "150 % " but got "1 % "
